US_22 terminada mais correcao de alguns testes

+++ app/Http/Controllers/MovementController.php
@@ -4,8 +4,10 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Movement;
+use App\Account;
 use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\DB;
 
 class MovementController extends Controller
 {
@@ -51,19 +53,42 @@ class MovementController extends Controller
             'description' => 'nullable',
             ]);
 
-        $lastMovementId = Movement::join('accounts', 'accounts.id','=', 'movements.account_id')
+        $movement['account_id'] = $account;
+        
+        $numMovements = Movement::where('account_id', '=', $account)
+                                ->select('movements.*')
+                                ->get();
+
+        $contaAtual = Account::findOrFail($account);
+
+        if (count($numMovements) == 0) {
+            $movement['start_balance'] = $contaAtual->start_balance;
+            
+            if ($request->type=='expense') {
+                $movement['end_balance'] = $contaAtual->start_balance - $request->value; 
+            }else{
+                $movement['end_balance'] = $contaAtual->start_balance + $request->value; 
+            }
+        }else{
+            $lastMovementId = Movement::join('accounts', 'accounts.id','=', 'movements.account_id')
                     ->where('accounts.id', '=', $account)
                     ->max('movements.id');
 
-        $lastMovement = Movement::findOrFail($lastMovementId);
+            $lastMovement = Movement::findOrFail($lastMovementId);
 
-        $movement['account_id'] = $account;
-        $movement['start_balance'] = $lastMovement->end_balance;
-        if ($request->type=='expense') {
-            $movement['end_balance'] = $lastMovement->end_balance - $request->value; 
-        }else{
-            $movement['end_balance'] = $lastMovement->end_balance + $request->value; 
+            
+            $movement['start_balance'] = $lastMovement->end_balance;
+            
+            if ($request->type=='expense') {
+                $movement['end_balance'] = $lastMovement->end_balance - $request->value; 
+            }else{
+                $movement['end_balance'] = $lastMovement->end_balance + $request->value; 
+            }
         }
+
+        Account::where('id', '=', $account)
+                ->update(['current_balance' => $movement['end_balance']]);
+        
         
         
         
@@ -82,17 +107,102 @@ class MovementController extends Controller
         if ($request->has('cancel')) {
             return redirect()->action('DashboardController@index', Auth::user());
         }
-
-        
         
         $movement = $request->validate([
             'type' => 'required',
             'date' => 'required',
             'value' => 'required',
             ]);
-    
 
         $movementModel = Movement::findOrFail($id);
+        
+        $numMovements = Movement::where('account_id', '=', $movementModel->account_id)
+                                ->select('movements.*')
+                                ->get();
+        
+        $lastMovementId = Movement::join('accounts', 'accounts.id','=', 'movements.account_id')
+                    ->where('accounts.id', '=', $movementModel->account_id)
+                    ->max('movements.id');
+
+        $lastMovement = Movement::findOrFail($lastMovementId);
+
+        $contaAtual = Account::findOrFail($movementModel->account_id);
+        
+        if (count($numMovements) == 0) {
+            if ($request->type=='expense') {
+                Movement::where('id', '=', $id)
+                    ->update(['start_balance' => $contaAtual->start_balance,
+                              'end_balance' => $contaAtual->start_balance - $request->value]);
+            }else{
+                Movement::where('id', '=', $id)
+                    ->update(['start_balance' => $contaAtual->start_balance,
+                              'end_balance' => $contaAtual->start_balance + $request->value]);
+            }
+
+        }  
+        
+        $movements = Movement::join('accounts', 'accounts.id','=', 'movements.account_id')
+                    ->where('accounts.id', '=', $movementModel->account_id)
+                    ->where('movements.id', '>=', $id)
+                    ->select('movements.*')
+                    ->orderBy('movements.id', 'asc')
+                    ->get();
+        //dd($movements[0]);
+        if (count($movements)>0) {
+            for ($i=0; $i < count($movements); $i++) {
+                if ($i==0) {
+                    if ($request->type=='expense') {
+                        Movement::where('id', '=', $id)
+                            ->update(['end_balance' => $movementModel->start_balance - $request->value]);
+                    }else{
+                        Movement::where('id', '=', $id)
+                            ->update(['end_balance' => $movementModel->start_balance + $request->value]);
+                    }
+                }else{
+                    $end_balance_movement = Account::join('movements', 'movements.account_id', '=', 'accounts.id')
+                            ->where('accounts.id', '=', $movementModel->account_id)
+                            ->where('movements.id', '=', $movements[($i-1)]->id)
+                            ->select('movements.end_balance')
+                            ->first();
+
+                 
+                    if ($movements[$i]->type == 'expense') {
+                        DB::table('accounts')
+                            ->join('movements', 'movements.account_id', '=', 'accounts.id')
+                            ->where('accounts.id', '=', $movementModel->account_id)
+                            ->where('movements.id', '=', $movements[$i]->id)
+                            ->update(['movements.start_balance' => $end_balance_movement->end_balance,
+                                    'movements.end_balance' => ($end_balance_movement->end_balance - $movements[$i]->value) ]);
+                        
+
+                    }else{
+                        DB::table('accounts')
+                            ->join('movements', 'movements.account_id', '=', 'accounts.id')
+                            ->where('accounts.id', '=', $movementModel->account_id)
+                            ->where('movements.id', '=', $movements[$i]->id)
+                            ->update(['movements.start_balance' => $end_balance_movement->end_balance,
+                                    'movements.end_balance' => ($end_balance_movement->end_balance + $movements[$i]->value)]);
+                    }
+                }   
+            }
+        }
+
+
+                
+        $lastMovementId = Movement::join('accounts', 'accounts.id','=', 'movements.account_id')
+        ->where('accounts.id', '=', $movement->account_id)
+        ->max('movements.id');
+
+        $lastMovement = Movement::findOrFail($lastMovementId);
+
+        DB::table('accounts')
+                ->where('accounts.id', '=', $movement->account_id)
+                ->update(['accounts.current_balance' => $lastMovement->end_balance]);
+        
+
+    
+
+        
         $movementModel->fill($movement);
         $movementModel->save();
             
@@ -102,7 +212,68 @@ class MovementController extends Controller
     public function delete($id)
     {
         $movement = Movement::find($id);
+
+        $account = Account::findOrFail($movement->account_id);
+
+        $numMovements = Movement::where('account_id', '=', $movement->account_id)
+                                ->select('movements.*')
+                                ->get();
+
+        if (count($numMovements)==1) {
+           DB::table('accounts')
+                ->where('accounts.id', '=', $movement->account_id)
+                ->update(['accounts.current_balance' => $account->start_balance]);
+        }
+
+
+        $movements = Movement::join('accounts', 'accounts.id','=', 'movements.account_id')
+                    ->where('accounts.id', '=', $movement->account_id)
+                    ->where('movements.id', '>=', $id)
+                    ->select('movements.*')
+                    ->orderBy('movements.id', 'asc')
+                    ->get();
+
+        if (count($movements) > 1) {
+            for ($i=1; $i < count($movements); $i++) {
+            $start_balance_movement = Account::join('movements', 'movements.account_id', '=', 'accounts.id')
+                            ->where('accounts.id', '=', $movement->account_id)
+                            ->where('movements.id', '=', $movements[($i-1)]->id)
+                            ->select('movements.start_balance')
+                            ->first();
+
+                if ($movements[$i]->type == 'expense') {
+                        DB::table('accounts')
+                            ->join('movements', 'movements.account_id', '=', 'accounts.id')
+                            ->where('accounts.id', '=', $movement->account_id)
+                            ->where('movements.id', '=', $movements[$i]->id)
+                            ->update(['movements.start_balance' => $start_balance_movement->start_balance,
+                                    'movements.end_balance' => ($start_balance_movement->start_balance - $movements[$i]->value) ]);
+                        
+
+                    }else{
+                        DB::table('accounts')
+                            ->join('movements', 'movements.account_id', '=', 'accounts.id')
+                            ->where('accounts.id', '=', $movement->account_id)
+                            ->where('movements.id', '=', $movements[$i]->id)
+                            ->update(['movements.start_balance' => $start_balance_movement->start_balance,
+                                    'movements.end_balance' => ($start_balance_movement->start_balance + $movements[$i]->value)]);
+                    }
+            }
+        }
+
         $movement->delete();
+
+        if (count($numMovements)>1) {
+            $lastMovementId = Movement::join('accounts', 'accounts.id','=', 'movements.account_id')
+            ->where('accounts.id', '=', $movement->account_id)
+            ->max('movements.id');
+
+            $lastMovement = Movement::findOrFail($lastMovementId);
+
+            DB::table('accounts')
+                    ->where('accounts.id', '=', $movement->account_id)
+                    ->update(['accounts.current_balance' => $lastMovement->end_balance]);
+        }
         
         return redirect()->action('DashboardController@index', Auth::user());
     }
